+++ apple-sdk/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 Released on ReleaseDate.
 
+* XROS support. `Platform` enumeration added `XrOs` and `XrOsSimulator`
+  variants. The `aarch64-apple-xros-sim` and `*-apple-xros` triples are
+  now recognized as XROS.
+
 ## 0.5.2
 
 Released on 2023-12-22.